requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
tensorflow
keras
numpy
opencv-python
pillow
pandas
reportlab
python-dateutil
Werkzeug
pyotp
qrcode
python-dotenv
beautifulsoup4
sib_api_v3_sdk
Flask-Migrate
pymysql
cryptography
gunicorn